Chunk upload fix for `multipart/form-data` form request.
We forced the `header` for the chunk upload request to be `multipart/form-data` but the header must be set auotmatically when sending the `FormData`. This was causing the backend being unable to read parameters sent

Changed the way we stop the chunk process when the upload fails in `upload` phase. We no longer set the `file.active = false` since that would prevent the next file to start uploading